moved buildModel to utils
