+++ lib/utils/index.js
@@ -1,4 +1,4 @@
-const { get, isEqual, mapValues, isPlainObject, omit, castArray } = require('lodash');
+const { get, isEqual, mapValues, isPlainObject, omit, castArray, reduce, isUndefined } = require('lodash');
 const { parse } = require('url');
 const qs = require('qs');
 const striptags = require('striptags');
@@ -75,6 +75,16 @@ const getSort = (column, state) => ({
   ascending: state.column === column ? !state.ascending : true
 });
 
+const buildModel = (...args) => {
+  return Object.assign(
+    ...args.map(schema => {
+      return reduce(schema, (fields, { nullValue }, key) => {
+        return { ...fields, [key]: isUndefined(nullValue) ? null : nullValue };
+      }, {});
+    })
+  );
+};
+
 module.exports = {
   toTitleCase,
   getTitle,
@@ -83,5 +93,6 @@ module.exports = {
   hasChanged,
   cleanModel,
   removeQueryParams,
-  queryStringFromState
+  queryStringFromState,
+  buildModel
 };
